+++ examples/ng-cli/src/app/app.module.ts
@@ -4,18 +4,18 @@ import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
-// import { Ng2FilterPipe } from 'ng2-filter-pipe';
-import { Ng2FilterPipe } from './shared/ng2-filter.pipe';
+import { Ng2FilterPipeModule } from 'ng2-filter-pipe';
+// import { Ng2FilterPipe } from './shared/ng2-filter.pipe';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    Ng2FilterPipe
+    AppComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule
+    HttpModule,
+    Ng2FilterPipeModule
   ],
   providers: [],
   bootstrap: [AppComponent]